renamed files having same names
