+++ packages/itwin/ec3-widget/src/components/EC3WidgetComponent.scss
